+++ components/lyric_matcher.py
@@ -1,5 +1,6 @@
 # Python
 import re
+import logging
 from dataclasses import dataclass
 from typing import List
 
@@ -7,6 +8,8 @@ from typing import List
 
 
 # 1st Party
+from .miscellaneous import percentage
+from lyric_aligner import WordAndTiming
 
 
 @dataclass
@@ -26,6 +29,20 @@ class MatchLyric:
     word_split_char_post: str   = " " # The pos-character that a lyric-word was split on. Most often a space " ", sometimes a hyphen "-"
     time_start: float           = -1.0
     time_end: float             = -1.0
+
+
+@dataclass
+class MatchResult:
+    words_total: int          = 0
+    words_matched: int        = 0
+    match_percentage: float   = -1.0
+
+    def __post_init__(self):
+        self.match_percentage = percentage(self.words_matched, self.words_total)
+
+    def get_string(self):
+        return f"{self.match_percentage :.2f}% ({self.words_matched} / {self.words_total})"
+
 
 
 class LyricMatcher():
@@ -294,3 +311,126 @@ class LyricMatcher():
 
     def _get_alignment_lyric_with_pre_post_chars(self, alignment_lyric):
         return alignment_lyric.word_original + alignment_lyric.word_split_char_post
+
+
+    def match_aligned_lyrics_with_structured_lyrics(self,
+        lyrics_time_aligned: list[WordAndTiming],
+        lyrics_structured: list[MatchLyric],
+        debug_print=False) -> List[MatchLyric]:
+        """ Matches time aligned lyrics with (sentance) structured lyrics.
+
+        Aligned lyrics and structured lyrics rarely match exactly. This function uses a moving
+        expanding/contracting window in order to find the best structured lyric, for each aligned
+        lyric.
+
+        The algorithm will assume that the time aligned lyrics is the more sparse data, and use
+        it for the main for-loop. As the structured lyrics are scanned for a match, a window is
+        used within which to do a forward-search. If a match isn't found within the window, the
+        time aligned lyric is skipped and the window expands. Every time a match if found, the
+        window shrinks.
+
+        Args:
+            - TODO:
+        Returns:
+            List[MatchLyric] - List of structure lyrics now with timing information
+
+        """
+
+        #debug_print = True
+
+        # We'd like to easily iterate through each structured lyric word, so we'll transform the
+        # the list of strings (in lyrics_structured) into a long list of recordtypes, each with
+        # a start / stop time, and a line
+
+        #lyrics_structured_better = self._convert_lyric_sentences_to_stuctured_words(lyrics_structured)
+
+        # Step 1. Process lyrics_structured to befit the incoming lyrics_timing.
+        # E.g. NUSAutoAlignLyrix removes ()'s and appears to perhaps not match "'em" so it should
+        # be replaced with 'them' (before processing though)
+
+        # Step 2. For each timed lyric, find the 'best' match
+
+        # Iterate though each timed lyric
+        # - Use fuzzy wuzzy to find 'match of word' vs word given in timing
+        #   - For now - expect exact match
+        #   - If a match is no-go, there should be 2 types of tolerances
+        #       1. When holding a timed lyric, how far can we skip through the structured to find a match? (start with 3)
+        #       2. Assuming the limit for 1. is hit (e.g. no match within the next 3 words), how many times
+        #       can we timed lyrices are we allowed to not match. (start with 3)
+        #   - Reset each of these as the work through the lyrics
+
+        # We assumed that timed lyrics may not precisely fit the structured lyrics, specifically:
+        # - A structured lyric word may not appear in the timed lyrics
+        #   - Either because the input lyrics to align were bunk, or
+        #   - because the lyric alignment algorithm failed to properly match the word
+        # - Timed lyric may include words that don't match a lyric
+        #   - NUSAutoAlignLyrix includes a 'Breath*' word that should either be
+        #       - Matched to a proper word the algorithm mistakenly thought was a breath
+        #       - Removed as we don't really want to display 'Breath*'...?
+         
+        # Matching the timed lyrics to structured lyrics is centered around the timed lyrics -
+        # i.e. we iterate over the timed lyrics as that is likely the more sparse data
+         
+        # wat -> word and timing
+        lsb_index = 0
+
+        total_unmatched = 0
+
+        # Growing / Shrinking tolerance window works well with BLUR - Boys & Girls
+        min_mismatch_tolerance = 3
+        current_mismatch_tolerance = min_mismatch_tolerance
+
+        for wat_index, wat in enumerate(lyrics_time_aligned):
+
+            failed_to_match = False
+
+            match_span = min(len(lyrics_structured) - lsb_index, current_mismatch_tolerance)
+            for index_offset in range(match_span):
+
+                if debug_print:
+                    self._debug_print(lyrics_time_aligned, wat_index, lyrics_structured, lsb_index, index_offset, current_mismatch_tolerance)
+                    input("Press Enter to continue...")
+
+                # Fix out of range...
+                lsb = lyrics_structured[lsb_index + index_offset]
+
+                #print(f"Matching {wat.word.lower()} against {lsb.word.lower()} ")
+
+                # Todo: Improve word-to-word comparisson 
+                if wat.word.lower() == lsb.word_alignment.lower():
+                    #lsb_mismatch_count = 0
+                    lyrics_structured[lsb_index + index_offset].time_start = wat.time_start
+                    lyrics_structured[lsb_index + index_offset].time_end = wat.time_end
+
+                    # If we find a 'later' match, we must move the lsb_index forward by the offset
+                    # otherwise, it'll keep falling behind
+                    lsb_index += index_offset
+                    
+                    break
+                else:
+                    #lsb_mismatch_count += 1
+                    horse = 456
+                
+                if index_offset == match_span-1:
+                    failed_to_match = True
+
+                #self._debug_print(lyrics_time_aligned, wat_index, lyrics_structured_better, lsb_index, index_offset, current_mismatch_tolerance)
+
+            if failed_to_match:
+                logging.debug(f"Failed to match a word: {wat.word:10} | wat_index: {wat_index:3} | lsb_index: {lsb_index:3}")
+                # Given that the timed lyrics are always expected to be more sparse, than
+                # the structured ones, if a match fails, it 'should' be ok to move the lsb_index
+                # ahead. See how we go re. this change.
+                #lsb_index += 1
+                current_mismatch_tolerance += 1
+                total_unmatched += 1
+            else:
+                current_mismatch_tolerance = max(min_mismatch_tolerance, current_mismatch_tolerance-1)
+                lsb_index += 1
+
+        num_time_aligned_lyrics = len(lyrics_time_aligned)
+        total_matched = num_time_aligned_lyrics - total_unmatched
+
+        match_result = MatchResult(num_time_aligned_lyrics, total_matched)
+
+        return lyrics_structured, match_result