+++ components/miscellaneous.py
@@ -1,0 +1,4 @@
+
+
+def percentage(numerator, denominator):
+    return 100 * float(numerator)/float(denominator) if denominator else 0