+++ lyric_aligner/lyric_aligner_disabled.py
@@ -1,4 +1,11 @@
-from lyric_aligner import LyricAlignerInterface
+# Python
+
+# 1st Party
+from .lyric_aligner_interface import LyricAlignerInterface
+from .lyric_aligner_interface import WordAndTiming
+
+# 3rd Party
+
 
 # Used to be named 'Off', but this is interpreted as 'False' by PyYaml
 class LyricAlignerDisabled(LyricAlignerInterface):
@@ -9,5 +16,5 @@ class LyricAlignerDisabled(LyricAlignerInterface):
     def _convert_to_wordandtiming(self, input):
         return []
 
-    def align_lyrics(self, path_to_audio_file, path_to_lyric_input, use_preexisting):
-        return ""
+    def align_lyrics(self, path_to_audio_file, path_to_lyric_input, use_preexisting) -> list[WordAndTiming]:
+        return []